Ensure token is properly initialized when using local credentials
When using a user's local credentials, the async client never reads the access token,
causing authentication failures when trying to connect to datastore. This works
when using service accounts. The difference is that for user credentials, Google's
auth library initializes the access token for you, but does not with service
accounts.

DatastoreImpl sets up authentication correctly only if the credential's access token
is  to be null, but for user credentials, the token will never be null.

This PR makes it so DatastoreImpl now handles this case correctly.